修改代码结构

+++ GateServer/GateServer.cpp
@@ -14,7 +14,7 @@ int main()
 	std::string gate_port_str = gCfgMgr["GateServer"]["Port"];
 	unsigned short gate_port = atoi(gate_port_str.c_str());
 	try {
-		unsigned short port = static_cast<unsigned short>(gate_port);
+		unsigned short port = gate_port;
 		net::io_context ioc{ 1 };  // 底层默认一个线程跑
 		boost::asio::signal_set signals(ioc, SIGINT, SIGTERM);
 		signals.async_wait([&ioc](const boost::system::error_code& error, int signal_errno) {
